requirements.txt
colorama>=0.4.6
pydantic>=2.4.2
typing-extensions>=4.5.0
aiohttp>=3.8.0
pillow>=9.0.0
numpy>=1.20.0
setuptools>=61.0.0
wheel>=0.37.0
twine>=4.0.0
